feat: Add new icons for close circle and check circle
This commit adds new SVG icons for close circle and check circle. The SVG files are added to the assets/icons directory. The icons are named 'close-circle.svg' and 'check-circle.svg' respectively. These changes enhance the project by providing new icon options for the 'closeCircle' and 'checkCircle' types in the Icon component.